Refactor code to create modules
