+++ src/index.js
@@ -1,45 +1,17 @@
-import './style.css';
+import "./style.css";
 
-import {
-  getMovies,
-  getMovieDetail,
-  getLikes,
-  getComments,
-  addLike,
-  addComment,
-} from './modules/store/API.js';
+import { addLike } from "./modules/store/API.js";
+import moviesDataCollection from "./modules/store/Data";
+import openPopup from "./modules/popup";
 
-const moviesList = document.querySelector('.movies');
-const popup = document.querySelector('.popup');
-const movies = getMovies();
-const likes = getLikes();
+const moviesList = document.querySelector(".movies");
 
-const moviesDataCollection = async (movies, likes) => {
-  const movieLikes = await likes;
-  let moviesData = await movies;
-  moviesData = moviesData.map((item) => {
-    const foundMovie = movieLikes.find(
-      (element) => Number(element.item_id) === Number(item.id),
-    ) || 0;
-    return {
-      id: item.id,
-      name: item.name,
-      image: item.image,
-      network: item.network.name,
-      country: item.network.country.name,
-      genres: item.genres,
-      status: item.status,
-      likes: foundMovie.likes || 0,
-    };
-  });
-  return moviesData;
-};
 const populateData = async (data) => {
   const moviesData = await data;
   moviesData.forEach((item) => {
-    const likeText = item.likes > 1 ? 'likes' : 'like';
-    const movieCard = document.createElement('div');
-    movieCard.className = 'movie-card';
+    const likeText = item.likes > 1 ? "likes" : "like";
+    const movieCard = document.createElement("div");
+    movieCard.className = "movie-card";
     movieCard.innerHTML = `
         <img src=${item.image.medium} alt="Movie pic"/>
         <div>
@@ -52,81 +24,18 @@ const populateData = async (data) => {
     moviesList.appendChild(movieCard);
   });
 };
-const openPopup = async (id) => {
-  let comments = await getComments(id);
-  const movie = await getMovieDetail(id);
-  if (comments.length) {
-    comments = comments.map(
-      (item) => `<li>${item.creation_date} ${item.username} : ${item.comment}</li>`,
-    );
-  }
 
-  popup.innerHTML = `
-  <div class="popup-content">
-  <div class="image-container">
-    <img
-      src=${movie.image.original}
-      alt="movie pic"
-    />
-    <button type="button" class="close-detail">
-      <i class="fa-solid fa-x"></i>
-    </button>
-  </div>
-  <h2>${movie.name}</h2>
-  <ul class="movie-info">
-    <li><span>Network</span> : ${movie.network.name}</li>
-    <li><span>County</span> : ${movie.network.country.name}</li>
-    <li><span>Genres</span> : ${movie.genres}</li>
-    <li><span>Status</span> : ${movie.status}</li>
-  </ul>
-  <h3>Comments(${comments.length})}</h3>
-  <ul class ="comments-container">
-   ${comments}
-  </ul>
-  <h3>Add Comment</h3>
-  <form action="" class="form">
-    <input type="text" placeholder="Your name" name= "userName"required />
-    <textarea name="commentText" id="" cols="30" rows="10" placeholder="Your Insights" required></textarea>
-    <button type="submit">Comment</button>
-  </form>
-  </div>
-  `;
-  popup.classList.remove('hide');
-
-  const commentsContainer = popup.querySelector('.comments-container');
-  const closeButton = popup.querySelector('.fa-x');
-  const commentForm = popup.querySelector('.form');
-  const { userName, commentText } = commentForm.elements;
-
-  closeButton.addEventListener('click', () => {
-    popup.classList.toggle('hide');
-  });
-
-  commentForm.addEventListener('submit', (e) => {
-    e.preventDefault();
-    addComment(id, userName.value, commentText.value);
-    const date = new Date();
-    const createdAt = `${date.getFullYear()}-${date.getMonth()}-${date.getDate()}`;
-    const commentLi = document.createElement('li');
-    commentLi.innerHTML = `${createdAt} ${userName.value} : ${commentText.value}`;
-    commentsContainer.appendChild(commentLi);
-  });
-};
-
-moviesList.addEventListener('click', (e) => {
-  if (e.target.className.includes('like-button')) {
+moviesList.addEventListener("click", (e) => {
+  if (e.target.className.includes("like-button")) {
     addLike(e.target.id);
     const likeCountSpan = document.querySelector(`#likes-${e.target.id}`);
     likeCountSpan.textContent = Number(likeCountSpan.textContent) + 1;
-    likeCountSpan.nextElementSibling.textContent = Number(likeCountSpan.textContent) > 1 ? 'likes' : 'like';
+    likeCountSpan.nextElementSibling.textContent =
+      Number(likeCountSpan.textContent) > 1 ? "likes" : "like";
   }
-  if (e.target.className === 'comment-button') {
+  if (e.target.className === "comment-button") {
     openPopup(e.target.id);
   }
 });
-// popup.addEventListener('click', (e) => {
-//   if (e.target.parentNode.className === 'close-detail') {
-//     popup.classList.toggle('hide');
-//   }
-// });
-populateData(moviesDataCollection(movies, likes));
+
+populateData(moviesDataCollection());
